server:IMPORTANT: replace callback functions by async/await functions
This is a main, huge change in the server, since all calls will be treated by using async (promise-like) instead of callbacking.
Async is easy to swallow for small functionality, but when the code starts co call async from everywhere, it is difficult not to make errors. This is due to all node core libreries are based on callacks and emmiters (on events), and async, by now, has to live with that kind of code.

While this model is harder(mentally) to implement than callbacks, the main advantage is ths simplicity for the end developer (the developer who uses the server). Now, the end developer can create his services with a lot of less code and the code will be easier to read (since it's like reading linear code, and not jumping forn callback to callback or wrapping everything into promises, etc...

This commit is not an stable commit (the server works, tests passed, but extensive tests have not been executed yet)