demo: new and nicer way of executing dynamic code by using async/await
The old version of test.js worked with callbacks. This one is purely asnyc. Lot of code in the future  won't suffer from callback hell and error management will be flawless. This test file is a fair demonstration on how easy can be develop apps without thinking on callbacks anymore.

Callbacks were not a bad thing. How people used callbacks (instead of splitting into functions and modules) was the bad thing.