demo:REST: cars service example by using async/await instead of callbacks
This is a deo on how to create REST services.

The new format has shinked the code by 2X, so the improvement on reading/writing REST services is HUGE.

This also demonstrates that the risk of changing from callbacks to async/await format is clearly justified later when developing web applications.