server:proxy: prevent proxying a local page with a referer header
The problem appeared with when chromium browser requested a favicon.ico after requesting a /test page. Since that second call contains referer, the server performed a proxy request, not a static request.

Not sure if the solution is OK for all proxy cases, but for now it is better to call proxy resources only whe they are directly called by the server. Security is first, always.