v8 harmony: performance tests when using async/await with node 7.0.0 --harmony
Perfomance has dropped near 10% (before changes), check if in the future this is solved. Native promises work better than native async so this should be v8 improvement later.

In the meantime, the test.js is replaced to sync function because in the express version of the code is also sync.