server: send '' when apps are not retunrnig data
The general fuunction for a request call is to return something.

If a developer forgets to return anything (if no payload return '' at least), the server will return '' automatically.

Example:

```javascript
function get(req, res, params) {
	// do whatever
	// return 'hello' // forgot to return something
}
```

The previous impl throws a cryptic error, now there is no error, if the payload return was missing, the developer will see that the response is void