response codes: return all the response codes the same way
before: if error throw(401) or similar was fine while response.statusCode=204 was used
now: all the status codes are managed the same way, no handy special shorttyping trick, if 20X set status, if 40X the same, if 50X the same.

Throw number is still available but in the documentation, the recommended way is tu use res.statusCode and return nothing or message.