benchmark: leave RAW code in server to measure barebone node vs functionality
Note: there is a raw server in benchmarks already.

The current change is just to show in the future how to start checking form raw code by removin functionalities until detecting the bottlenecks of existing or new functioonalities.

for instance, at this time, a raw node server produces 19K/req while a1-server is only 11.5. So it is ok to check what functionalities are delaying the app. 19K/req is unreachable (routing, check static/dynamic, etc) but throughput could be better.