router: put `params == null` when REST services
This was a bug because since version 3.0.X  params can be null if no REST or querystring is in the url.

For normal pages it was ok but for REST services (marked in the "rules" rpoerpty, the match() method leads to `{}`. It is better to return everything to null (despite of losing functional philosophy
, because it is easier in all the dynamic pages to check `if (!params) then whatever`