security: do not the real error details to the front end
For instance if error is not castrated, the end user could receive ENOENT file a/b/c/secrets.config was not found. That is fine when developing but very bad idea on production.

So for security reasons, the exceptions are purified, and the developer should check in the log server, but end users cannot get good info.

If a service has to send error data, better send a valid response with status code 400 500 etc  and descripton, not throwing Exceptions naively.