security: the response status and text has standard HTTP text.
Old format: no status code and custom general text.
new format: message with status code and generic http status description.

This new format is as secure as the old one but this is handier for developers since th http status code is read directly in the body payload instead of looking at headers.