server: log real error if require(app/file) fails
require(file) can throw error when file not found but also when the js code itself contains runtime errors when loading.

The real error is not sent to the user (security) due to the built-in error cleaning. But for developers it is important to know that the proble is that the js file is not correct.