proxy: MAJOR, re-assess and simplify proxy
The initial change was to adapt proxy for accepting querystring and hashes that were not implemented before. But when reviewing the code it was a bit messy and stil do not understand why referer was implemented. If in the future is necessary, it is probably better to add `origin` header than `referer`( see MDN referer or wikipedia to see the list of privacy problems with referer.