proxy: IMPORTANTadd a pure http to avoid 301 redirections
Just to avoid potential misunderstanding.

For example http:google.com sends a 301 and redirects to https://. The problem is that if we use http as proxyrule , when we test localhost/google, the address bar in browser do not shw that, but the final https://google. This may cause confusion thinking that the reverse proxy is not working fine, while, in fact it is working. The best approach is to add a pure http example page, so in case we have the problem again, we can compare against tests, showing that the reverse proxy works OK.