MAJOR: support for ESM modules
This commit is an important step towards migration (not only this project) to ESM. It should not break compatibility (so it will be marked as SEMVER MINOR), only an small performance loss probably.

While core is still CJS, this change makes that we can create from now new projects with ESM module system, since it is the accepted path also for the node community.