server: allow starting different port as "node index 8081"
In some cases, the PORT=8081 as env variable is not possible. This new option adds flexibility to run different ports for dev and prod, while the custom parsing of args remains intact because the developer can use the custom config object server.start(conf) in case the first arg for the applications is also a number but is not related to the port.